prima implementazione

+++ src/nnue/nnue_type.h
@@ -18,14 +18,14 @@
 #ifndef NNUE_TYPE_H_
 #define NNUE_TYPE_H_ 
 
-using flOutType = int16_t;
-using flAccumulatorType = int16_t;
-using flBiasType = int16_t;
-using flWeightType = int16_t;
+using flOutType = int32_t;
+using flAccumulatorType = int32_t;
+using flBiasType = int32_t;
+using flWeightType = int32_t;
 
-using outType = int16_t;
+using outType = int32_t;
 using accumulatorType = int32_t;
 using biasType = int32_t;
-using weightType = int8_t;
+using weightType = int32_t;
 
 #endif
